Refactoring and added test for ingestion source
